Accion borrar usuarios

+++ src/model/usersModel.js
@@ -1,5 +1,6 @@
 const fs = require("fs");
 const path = require("path");
+const bcryptjs = require("bcryptjs");
 
 const usersModel = {
   getAll: function () {
@@ -18,60 +19,63 @@ const usersModel = {
   },
 
   //  update: function (req) {
-  //   let productList = this.getAll();
-  //   let foundProduct = this.getOne(req.params.id);
-  //   let updatedProduct = req.body;
-  //   updatedProduct.id = foundProduct.id;
-  //   updatedProduct.image = foundProduct.image;
-  //   updatedProduct.alt = foundProduct.name;
-  //   updatedProduct.price = parseFloat(updatedProduct.price);
-  //   updatedProduct.stock = parseFloat(updatedProduct.stock);
+  //   let usersList = this.getAll();
+  //   let foundusers = this.getOne(req.params.id);
+  //   let updatedusers = req.body;
+  //   updatedusers.id = foundusers.id;
+  //   updatedusers.image = foundusers.image;
+  //   updatedusers.alt = foundusers.name;
+  //   updatedusers.price = parseFloat(updatedusers.price);
+  //   updatedusers.stock = parseFloat(updatedusers.stock);
 
-  //   let modifyProducts = productList.map((product) => {
-  //     if (product.id == updatedProduct.id) {
-  //       return updatedProduct;
+  //   let modifyuserss = usersList.map((users) => {
+  //     if (users.id == updatedusers.id) {
+  //       return updatedusers;
   //     } else {
-  //       return product;
+  //       return users;
   //     }
   //   });
   //   if (req.file) {
-  //     if (foundProduct.image) {
-  //       fs.unlinkSync(`${__dirname}/../../public/images/products/${foundProduct.image}`);
+  //     if (foundusers.image) {
+  //       fs.unlinkSync(`${__dirname}/../../public/images/userss/${foundusers.image}`);
   //     }
-  //     updatedProduct.image = `imagen - ${path.basename(req.file.originalname)}`;
+  //     updatedusers.image = `imagen - ${path.basename(req.file.originalname)}`;
   //   }
-  //   this.writeFile(modifyProducts);
+  //   this.writeFile(modifyuserss);
   // },
 
   store: function (req) {
     let newUser = req.body;
     newUser.id = this.lastID() + 1;
-    newUser.alt = req.body.firtName;
-    newUser.image = "imagen - perfil " + path.basename(req.file.originalname);
+    newUser.alt = req.body.firstName;
+    if(req.file){newUser.image = "imagen - perfil " + path.basename(req.file.originalname)};
     newUser.admin = false;
-
+    if(newUser.password !== newUser.confirmPassword){
+      return false
+    }
+    newUser.password = bcryptjs.hashSync(newUser.password, 10);
+    delete newUser.confirmPassword;
     let newUserList = this.getAll();
     newUserList.push(newUser);
-
     this.writeFile(newUserList);
   },
 
-  // delete: function (req) {
-  //   let productList = this.getAll();
-  //   let indexProduct = req.params.id;
-  //   let deletedProduct = productList.filter((product) => {
-  //     if (product.id != indexProduct) {
-  //       return product;
-  //     }
-  //   });
+  delete: function (req) {
+    let usersList = this.getAll();
+    let indexUsers = req.params.id;
+    let deletedUsers = usersList.filter((users) => {
+      if (users.id != indexUsers) {
+        return users;
+      }
+    });
 
-  //   let productToDelete = this.getOne(indexProduct);
-  //   if (productToDelete.image) {
-  //     fs.unlinkSync(`${__dirname}/../../public/images/products/${productToDelete.image}`);
-  //   }
+    let usersToDelete = this.getOne(indexUsers);
+    if (usersToDelete.image) {
+      fs.unlinkSync(`${__dirname}/../../public/images/users/${usersToDelete.image}`);
+    }
 
-  //   this.writeFile(deletedProduct);
-  // },
+    this.writeFile(deletedUsers);
+  },
 
   lastID: function () {
     let usersData = this.getAll();
